[IE-110] 다국어 디렉토리 선택 UI 변경
