--- apps/client/src/components/DirectorySelector/components/FileManager/components/FileManagerHeader/components/FileManagerSearch/components/FileManagerExitSearchIconButton/index.tsx
@@ -1,25 +1,0 @@
-import { useFileManagerStore } from '@components/DirectorySelector/stores/fileManagerStore';
-import CloseIcon from '@mui/icons-material/Close';
-import { IconButton, Tooltip } from '@mui/material';
-
-function FileManagerExitSearchIconButton() {
-  const handleClick = useFileManagerStore(({ finishSearch }) => finishSearch);
-
-  return (
-    <Tooltip title={'검색창 닫기 (esc)'}>
-      <IconButton
-        aria-label={'exit search'}
-        size={'small'}
-        onClick={handleClick}
-        sx={{
-          position: 'absolute',
-          right: 0,
-        }}
-      >
-        <CloseIcon fontSize={'medium'} />
-      </IconButton>
-    </Tooltip>
-  );
-}
-
-export default FileManagerExitSearchIconButton;
