+++ apps/client/src/components/FileManagerDialog/components/FileManagerHeader/components/FileManagerSearch/components/FileManagerExitSearchIconButton/index.tsx
@@ -1,0 +1,25 @@
+import { useFileManagerDialogStore } from '@components/FileManagerDialog/stores';
+import CloseIcon from '@mui/icons-material/Close';
+import { IconButton, Tooltip } from '@mui/material';
+
+function FileManagerExitSearchIconButton() {
+  const handleClick = useFileManagerDialogStore(({ finishSearch }) => finishSearch);
+
+  return (
+    <Tooltip title={'검색창 닫기 (esc)'}>
+      <IconButton
+        aria-label={'exit search'}
+        size={'small'}
+        onClick={handleClick}
+        sx={{
+          position: 'absolute',
+          right: 0,
+        }}
+      >
+        <CloseIcon fontSize={'medium'} />
+      </IconButton>
+    </Tooltip>
+  );
+}
+
+export default FileManagerExitSearchIconButton;
