+++ apps/client/src/components/FileManagerDialog/components/FileManagerHeader/components/FileManagerSearch/defines/animation.ts
