+++ apps/client/src/components/FileManagerDialog/hooks/useFileManagerDialogCloseHandler.ts
@@ -1,0 +1,8 @@
+import { useFileManagerDialogStore } from '@components/FileManagerDialog/stores';
+
+function useFileManagerDialogCloseHandler() {
+  const close = useFileManagerDialogStore(({ close }) => close);
+  return () => close(({ viewType, directoryOnly }) => ({ viewType, directoryOnly }));
+}
+
+export default useFileManagerDialogCloseHandler;
