+++ apps/client/src/components/Layout/components/Header/index.tsx
@@ -1,3 +1,4 @@
+import LocaleDirectoryAddButton from '@components/Layout/components/Header/components/LocaleDirectoryAddButton';
 import SidebarController from '@components/Layout/components/Header/components/SidebarController';
 import { AppBar, Toolbar, Typography, useTheme } from '@mui/material';
 
@@ -13,6 +14,7 @@ function Header() {
         <Typography variant={'h6'} noWrap component={'div'} sx={{ flexGrow: 1, display: { xs: 'none', sm: 'block' }, padding: '0 8px' }}>
           i18n Editor
         </Typography>
+        <LocaleDirectoryAddButton />
       </Toolbar>
     </AppBar>
   );
